+++ crates/tabby-inference/src/lib.rs
@@ -24,6 +24,17 @@ pub struct TextGenerationOptions {
     pub language: Option<&'static Language>,
 }
 
+impl TextGenerationOptions {
+    pub const DEFAULT_TEMPERATURE: f32 = 0.1;
+
+    pub fn default_seed() -> u64 {
+        std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap()
+            .as_millis() as u64
+    }
+}
+
 #[async_trait]
 pub trait TextGeneration: Sync + Send {
     async fn generate(&self, prompt: &str, options: TextGenerationOptions) -> String;
