+++ crates/tabby/src/services/completion.rs
@@ -17,10 +17,7 @@ use tracing::debug;
 use utoipa::ToSchema;
 
 use super::model;
-use crate::{
-    services::{default_seed, DEFAULT_TEMPERATURE},
-    Device,
-};
+use crate::Device;
 
 #[derive(Error, Debug)]
 pub enum CompletionError {
@@ -232,8 +229,12 @@ impl CompletionService {
         let language = request.language_or_unknown();
         let options = Self::text_generation_options(
             language.as_str(),
-            request.temperature.unwrap_or(DEFAULT_TEMPERATURE),
-            request.seed.unwrap_or_else(default_seed),
+            request
+                .temperature
+                .unwrap_or(TextGenerationOptions::DEFAULT_TEMPERATURE),
+            request
+                .seed
+                .unwrap_or_else(TextGenerationOptions::default_seed),
         );
 
         let (prompt, segments, snippets) = if let Some(prompt) = request.raw_prompt() {
